Deleted restricted middleware and renamed firebaseDecoder because they were virtually the same.
