+++ server/routes/root_router.js
@@ -7,8 +7,7 @@
 
 const express = require('express');
 const RootController = require('../controllers/RootController');
-const decodemiddleware= require('../middleware/firebasedecoder');
-const restrictedMiddleware = require('../middleware/restricted');
+const restrictedmiddleware= require('../middleware/restricted');
 
 /**
  * Define router
@@ -29,7 +28,7 @@ router.route('/users')
   .get(RootController.findUsers)
 
 router.route('/users/auth')
-  .all(decodemiddleware)
+  .all(restrictedmiddleware)
   .post(RootController.auth)
 
 /**
